Add Web Feed auto-detect, About section, and split feature files
This commit folds together three related changes that touch most
of the Petal codebase, because they all land on the same files:

1. **Auto-detect selectors.**  A new `PetalAutoDetect` helper
   runs a heuristic pattern finder over the fetched HTML and
   hands back a seeded recipe.  It walks every `<a href>` on the
   page, fingerprints each link's 3-ancestor container path, and
   scores fingerprints by how many distinct anchors they cover,
   preferring fingerprints with a class over bare tag names.
   Inside the first matching container it probes common selectors
   for title / link / image / date / summary.  The builder
   surfaces this as a **Wand** button inside the selectors
   section; it's gated on the page having already been fetched so
   there's real HTML to analyse.

2. **"About" section in the management view.**  Mirrors the
   layout of `ClearThisPageSettingsView` so the entry point
   documents what Web Feeds is before diving into the imported
   list.  (The English copy lives alongside the existing `Petal.*`
   strings; translations for the other eight supported locales
   still need to be filled in.)

3. **File split.**  Three monoliths are reorganised into
   responsibility-scoped folders so each file fits on one screen:

   - `Shared/Petal/PetalZip.swift` → `PetalZip/` with
     `PetalZip`, `+Writer`, `+Reader`, `+CRC32`, and
     `Data+LittleEndian`.
   - `Shared/Petal/PetalPackage.swift` → `PetalPackage/` with
     `PetalPackage`, `+Export`, `+Import`.
   - `Shared/Petal/PetalEngine.swift` → `PetalEngine/` with
     `PetalEngine`, `+Fetching`, `+Parsing`, `+Dates`.
   - `PetalBuilderView.swift` stops using `@ViewBuilder private`
     helpers and extracts the three form sections into
     `PetalBuilderSourceSection`, `PetalBuilderSelectorsSection`,
     and `PetalBuilderPreviewSection`, plus a shared
     `PetalSelectorField` row.

   `project.pbxproj` picks up the new file paths in its Add Feed
   / WidgetsExtension exclusion sets so the SwiftSoup-and-
   WebKit-only files still don't bleed into those targets.

   The Info.plist UTType description is renamed from "Sakura RSS
   Package" to "Sakura Web Feed" so the Files app displays the
   product-facing name.
