+++ SakuraRSS/Views/Petal/PetalBuilderView.swift
@@ -1,22 +1,26 @@
 import SwiftUI
 
-/// Selector-based recipe editor for a single Petal.
+/// Selector-based recipe editor for a single Web Feed.
 ///
-/// The view keeps things approachable for users who've never heard of
-/// CSS selectors:
+/// The view keeps things approachable for users who've never
+/// heard of CSS selectors:
 ///
 /// 1. They paste a URL and tap **Fetch**.
-/// 2. The builder loads the page and shows a live preview of every
-///    item that currently matches `itemSelector`.
-/// 3. Each selector field has an info button explaining what it does
-///    and a sensible default.
+/// 2. The builder loads the page and shows a live preview of
+///    every item that currently matches `itemSelector`.
+/// 3. **Auto-Detect** runs a heuristic pattern finder that tries
+///    to guess the right selectors for them.
 /// 4. Saving calls `FeedManager.addPetalFeed` (create) or
 ///    `updatePetalRecipe` (edit), which re-routes the refresh to
 ///    `PetalEngine`.
 ///
-/// The preview re-runs on a debounced basis so heavy typing doesn't
-/// hammer the network; fetches are cached so tweaking selectors
-/// after an initial fetch is instantaneous.
+/// The preview re-runs on a debounced basis so heavy typing
+/// doesn't hammer the network; fetches are cached so tweaking
+/// selectors after an initial fetch is instantaneous.
+///
+/// This view owns the state and action methods; the three form
+/// sections (source, selectors, preview) live in their own
+/// sibling files so each is small enough to read in one screen.
 struct PetalBuilderView: View {
 
     enum Mode {
@@ -54,11 +58,31 @@ struct PetalBuilderView: View {
     var body: some View {
         NavigationStack {
             Form {
-                sourceSection
-                selectorsSection
-                previewSection
+                PetalBuilderSourceSection(
+                    name: $recipe.name,
+                    siteURL: $recipe.siteURL,
+                    fetchMode: $recipe.fetchMode,
+                    isFetching: isFetching,
+                    onFetch: { Task { await fetchAndPreview(force: true) } }
+                )
+                PetalBuilderSelectorsSection(
+                    recipe: $recipe,
+                    canAutoDetect: fetchedHTML != nil && !isFetching,
+                    onAutoDetect: runAutoDetect,
+                    onSelectorChanged: schedulePreview
+                )
+                PetalBuilderPreviewSection(
+                    articles: previewArticles,
+                    errorMessage: errorMessage,
+                    isFetching: isFetching,
+                    hasFetchedHTML: fetchedHTML != nil
+                )
                 if case .edit = mode {
-                    deleteSection
+                    Section {
+                        Button("Petal.Builder.Delete", role: .destructive) {
+                            showDeleteConfirm = true
+                        }
+                    }
                 }
             }
             .animation(.smooth.speed(2.0), value: previewArticles.count)
@@ -84,9 +108,7 @@ struct PetalBuilderView: View {
                 Text("Petal.Builder.DeleteConfirm.Message")
             }
             .onAppear {
-                if case .create = mode, !recipe.siteURL.isEmpty {
-                    Task { await fetchAndPreview() }
-                } else if case .edit = mode, !recipe.siteURL.isEmpty {
+                if !recipe.siteURL.isEmpty {
                     Task { await fetchAndPreview() }
                 }
             }
@@ -96,175 +118,7 @@ struct PetalBuilderView: View {
         }
     }
 
-    // MARK: - Sections
-
-    @ViewBuilder
-    private var sourceSection: some View {
-        Section {
-            TextField("Petal.Builder.Name.Placeholder", text: $recipe.name)
-                .textInputAutocapitalization(.words)
-            TextField("Petal.Builder.URL.Placeholder", text: $recipe.siteURL)
-                .keyboardType(.URL)
-                .textContentType(.URL)
-                .autocorrectionDisabled()
-                .textInputAutocapitalization(.never)
-            Picker("Petal.Builder.FetchMode", selection: $recipe.fetchMode) {
-                Text("Petal.Builder.FetchMode.Static")
-                    .tag(PetalRecipe.FetchMode.staticHTML)
-                Text("Petal.Builder.FetchMode.Rendered")
-                    .tag(PetalRecipe.FetchMode.rendered)
-            }
-            Button {
-                Task { await fetchAndPreview(force: true) }
-            } label: {
-                HStack {
-                    Text("Petal.Builder.Fetch")
-                    if isFetching {
-                        Spacer()
-                        ProgressView()
-                    }
-                }
-            }
-            .disabled(recipe.siteURL.isEmpty || isFetching)
-        } header: {
-            Text("Petal.Builder.Section.Source")
-        } footer: {
-            Text("Petal.Builder.Section.SourceFooter")
-        }
-    }
-
-    @ViewBuilder
-    private var selectorsSection: some View {
-        Section {
-            selectorField(
-                label: "Petal.Builder.ItemSelector",
-                text: $recipe.itemSelector,
-                placeholder: "article, li.post, [data-testid=card]"
-            )
-            selectorField(
-                label: "Petal.Builder.TitleSelector",
-                text: Binding(
-                    get: { recipe.titleSelector ?? "" },
-                    set: { recipe.titleSelector = $0.isEmpty ? nil : $0 }
-                ),
-                placeholder: "h2, .title"
-            )
-            selectorField(
-                label: "Petal.Builder.LinkSelector",
-                text: Binding(
-                    get: { recipe.linkSelector ?? "" },
-                    set: { recipe.linkSelector = $0.isEmpty ? nil : $0 }
-                ),
-                placeholder: "a, a.post-link"
-            )
-            selectorField(
-                label: "Petal.Builder.SummarySelector",
-                text: Binding(
-                    get: { recipe.summarySelector ?? "" },
-                    set: { recipe.summarySelector = $0.isEmpty ? nil : $0 }
-                ),
-                placeholder: "p.excerpt, .summary"
-            )
-            selectorField(
-                label: "Petal.Builder.ImageSelector",
-                text: Binding(
-                    get: { recipe.imageSelector ?? "" },
-                    set: { recipe.imageSelector = $0.isEmpty ? nil : $0 }
-                ),
-                placeholder: "img, .hero img"
-            )
-            selectorField(
-                label: "Petal.Builder.DateSelector",
-                text: Binding(
-                    get: { recipe.dateSelector ?? "" },
-                    set: { recipe.dateSelector = $0.isEmpty ? nil : $0 }
-                ),
-                placeholder: "time, .published"
-            )
-        } header: {
-            Text("Petal.Builder.Section.Selectors")
-        } footer: {
-            Text("Petal.Builder.Section.SelectorsFooter")
-        }
-        .onChange(of: recipe.itemSelector) { _, _ in schedulePreview() }
-        .onChange(of: recipe.titleSelector) { _, _ in schedulePreview() }
-        .onChange(of: recipe.linkSelector) { _, _ in schedulePreview() }
-        .onChange(of: recipe.summarySelector) { _, _ in schedulePreview() }
-        .onChange(of: recipe.imageSelector) { _, _ in schedulePreview() }
-        .onChange(of: recipe.dateSelector) { _, _ in schedulePreview() }
-    }
-
-    @ViewBuilder
-    private var previewSection: some View {
-        Section {
-            if let errorMessage {
-                Text(errorMessage)
-                    .foregroundStyle(.red)
-            } else if previewArticles.isEmpty && !isFetching && fetchedHTML != nil {
-                Text("Petal.Builder.Preview.Empty")
-                    .foregroundStyle(.secondary)
-            } else if previewArticles.isEmpty && fetchedHTML == nil {
-                Text("Petal.Builder.Preview.TapFetch")
-                    .foregroundStyle(.secondary)
-            } else {
-                ForEach(previewArticles.prefix(20), id: \.url) { article in
-                    VStack(alignment: .leading, spacing: 4) {
-                        Text(article.title)
-                            .font(.body)
-                            .lineLimit(2)
-                        Text(article.url)
-                            .font(.caption)
-                            .foregroundStyle(.secondary)
-                            .lineLimit(1)
-                        if let summary = article.summary, !summary.isEmpty {
-                            Text(summary)
-                                .font(.caption)
-                                .foregroundStyle(.secondary)
-                                .lineLimit(2)
-                        }
-                    }
-                    .padding(.vertical, 2)
-                }
-            }
-        } header: {
-            HStack {
-                Text("Petal.Builder.Section.Preview")
-                if !previewArticles.isEmpty {
-                    Spacer()
-                    Text("\(previewArticles.count)")
-                        .foregroundStyle(.secondary)
-                        .textCase(nil)
-                }
-            }
-        }
-    }
-
-    @ViewBuilder
-    private var deleteSection: some View {
-        Section {
-            Button("Petal.Builder.Delete", role: .destructive) {
-                showDeleteConfirm = true
-            }
-        }
-    }
-
-    // MARK: - Helpers
-
-    private func selectorField(
-        label: LocalizedStringKey,
-        text: Binding<String>,
-        placeholder: String
-    ) -> some View {
-        VStack(alignment: .leading, spacing: 4) {
-            Text(label)
-                .font(.caption)
-                .foregroundStyle(.secondary)
-            TextField(placeholder, text: text)
-                .font(.system(.body, design: .monospaced))
-                .autocorrectionDisabled()
-                .textInputAutocapitalization(.never)
-        }
-    }
+    // MARK: - Derived state
 
     private var canSave: Bool {
         !recipe.name.trimmingCharacters(in: .whitespaces).isEmpty
@@ -308,6 +162,33 @@ struct PetalBuilderView: View {
         previewArticles = PetalEngine.parse(html: html, recipe: recipe)
         errorMessage = previewArticles.isEmpty
             ? String(localized: "Petal.Error.NoMatches") : nil
+    }
+
+    /// Runs the heuristic selector finder against the currently
+    /// cached HTML and folds its suggestions into the editable
+    /// recipe.  Preserves any fields the user has already filled
+    /// in (notably `name`, which is usually already set by the
+    /// AddFeed flow that seeds the builder with a URL).
+    private func runAutoDetect() {
+        guard let html = fetchedHTML else { return }
+        guard let suggestion = PetalAutoDetect.detect(
+            html: html, siteURL: recipe.siteURL
+        ) else {
+            errorMessage = String(localized: "WebFeed.Builder.AutoDetect.Failed")
+            return
+        }
+        if recipe.name.isEmpty {
+            recipe.name = suggestion.name
+        }
+        recipe.itemSelector = suggestion.itemSelector
+        recipe.titleSelector = suggestion.titleSelector
+        recipe.linkSelector = suggestion.linkSelector
+        recipe.summarySelector = suggestion.summarySelector
+        recipe.imageSelector = suggestion.imageSelector
+        recipe.dateSelector = suggestion.dateSelector
+        recipe.dateAttribute = suggestion.dateAttribute
+        errorMessage = nil
+        runPreviewFromCache()
     }
 
     // MARK: - Save / Delete
